+++ dev/templates/nginx/.nsmignore
@@ -1,0 +1,2 @@
+# Define list of ignored files & directories to NOT include in the container,
+# as you would do in .gitignore
