+++ dev/templates/nginx/settings.yml
@@ -1,0 +1,12 @@
+name: "Nginx"
+description: "Nginx template"
+port_range:
+  min: 22222
+  max: 33333
+defaults:
+  ram: 512000000 # bytes
+  cpu: 2 # cores
+  disk: 2000000000 # bytes
+meta: {}
+# Optional ENV vars, and their default values
+env: {}
