+++ openapi.yml
@@ -186,6 +186,10 @@ components:
           type: object
           required: false
           description: "A map of custom variables mapped to values whose definitions are in settings.yml in template under 'env'."
+        meta:
+          type: object
+          required: false
+          description: "A map of custom optional variables. string -> string"
 paths:
   /v1/status:
     get:
@@ -388,46 +392,6 @@ paths:
                 $ref: "#/components/schemas/Result"
         "409":
           description: "Service is not running"
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Result"
-  /v1/service/{serviceId}/stopcmd:
-    post:
-      description: "Stop a service using cmd"
-      parameters:
-        - name: "serviceId"
-          in: "path"
-          required: true
-          schema:
-            type: "string"
-      responses:
-        "200":
-          description: "Stop command successfully sent."
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Result"
-        "400":
-          description: "Invalid request"
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Result"
-        "404":
-          description: "Service not found"
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Result"
-        "409":
-          description: "Conflict, service is not running"
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Result"
-        "500":
-          description: "Internal server error"
           content:
             application/json:
               schema:
